+++ src/commands/start.command.ts
@@ -1,6 +1,7 @@
+import { StartCommandOptions } from "../daemon-command/start.deamon-command";
 import { connect } from "./connect";
 
-export const start = async (names: string[]): Promise<void> => {
+export const start = async (options: StartCommandOptions): Promise<void> => {
     const client = await connect();
-    client.write(`start ${JSON.stringify(names)}`);
+    client.write(`start ${JSON.stringify(options)}`);
 };