+++ src/daemon-command/logs.daemon-command.ts
@@ -16,12 +16,6 @@ export function logsDaemonCommand(daemon: Daemon, socket: Socket, names: string[
             socket.write(`${script.name}: ${line}\n`);
         }
 
-        script.logSockets.push(socket);
-        socket.on("close", () => {
-            const index = script.logSockets.findIndex((i) => i == socket);
-            if (index !== -1) {
-                script.logSockets.splice(index, 1);
-            }
-        });
+        script.addLogSocket(socket);
     }
 }